% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghostr.R
\docType{data}
\name{ghost_sightings}
\alias{ghost_sightings}
\title{Ghost sightings in Kentucky}
\format{A data frame with 294 rows and 6 variables:
\describe{
  \item{url}{The URL to a Ghosts of America page containing more information about the sighting(s).}
  \item{city}{The city of the sighting(s).}
  \item{state}{The state of the sighting(s).}
  \item{sightings}{The number of sightings in that city.}
  \item{lat}{The latitude of the city.}
  \item{lon}{The longitude of the city}
}}
\usage{
data("ghost_sightings")
}
\description{
a dataset of ghost sightings in the state of Kentucky in the
United States.
}
\keyword{datasets}

