% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghostr.R
\docType{package}
\name{ghostr}
\alias{ghostr}
\alias{ghostr-package}
\title{A data package of ghost sightings}
\description{
The \code{ghostr} package provides a dataset of
ghost sightings in the state of Kentucky, \code{\link{ghost_sightings}},
for use in state-level mapping projects and examples.
}

